% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{blockfile}
\alias{blockfile}
\title{Function creates a block file to be read by dss.64.c.exe}
\usage{
blockfile(rateobj, gridimage, NAval = -999)
}
\arguments{
\item{rateobj, }{string, name of list, output of function irates()}

\item{gridimage, }{string,  name of block file, a SpatialPixelsDataFrame object}

\item{NAval, }{numeric, integer with grid value for "No data"}
}
\value{
\item{gridpars}{list parameters from block data: number rows and columns, resolution, (x,y), coordinates of the origin, NA value}
\item{outgrid}{list vector node values, vector id of regions, and number of regions   }
\item{file}{list characters; indicating the day of data collection, filename and root folder where file is stored}
\item{ingrid}{the SpatialPixelsDataFrame object used}
}
\description{
Converts a SpatialPixelsDataFrame object representing block data into a text file (.out) that is stored in `input` folder.
As input you should provide a SpatialPixelsDataFrame object name representing block data, with id region values at simulation grid nodes,
and a list returned by funtion irates().
}
\details{
The gridded values should refer to the region id's at simulation locations (nodes). Every regions recorded in disease data file should be assigned to 1 or more grid node.
}
