% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{irates}
\alias{irates}
\title{Creates an disease incidence rates file to be read by dss.64.c.exe}
\usage{
irates(
  dfobj = NA,
  oid = NA,
  xx = NA,
  yy = NA,
  zz = NA,
  cases = NA,
  pop = NA,
  casesNA = 2,
  day = NA,
  perhab = 10^5
)
}
\arguments{
\item{dfobj}{string, dataframe name with disease data}

\item{oid}{character, fieldname for region id}

\item{xx}{character, fieldname for x-coordinates}

\item{yy}{character, fieldname for y-coordinates}

\item{zz}{character, fieldname for z-coordinates}

\item{cases}{character, fieldname for number of cases}

\item{pop}{character, fieldname for population size}

\item{casesNA, }{numeric, an integer used to replace rows with cases = NA}

\item{day}{character, string indicating the date (format 'yyyymmdd') of disease cases}

\item{perhab}{numeric, an integer indicating how much the rate is multiplied for, to express the disease rate (e.g 10000 or 100000)}
}
\value{
The following list of objects:
\item{rates}{dataframe; results containing id region, x,y,z coordinates, incidence rate, variance-error term and population by region}
\item{mrisk}{numeric; estimated global risk (/`perhab`)}
\item{file}{list characters; indicating the day of data collection, filename and root folder where text file is stored }
\item{ssdpars}{list numeric; list of values to be passed to `ssdpars()`}
}
\description{
Function computes disease incidence rates and variance-error terms by region,
and writes the result into a text file (.out) that is stored in `./input` folder.
As input you should provide a dataframe with disease cases, population size
and cartesian coordinates by region. Columns should include the following data:
id of region, x, y and z cartesian coordinates at regions mass center,
number of disease cases and population size by region.
}
\details{
The function also writes and stores a text file (.out). This file contains x, y, z and incidence rate by row (region) using GeoEAS file format.
}
