% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{maskfile}
\alias{maskfile}
\title{Function creates a mask file to be read by dss.64.c.exe}
\usage{
maskfile(blockobj)
}
\arguments{
\item{blockobj, }{string, name of list, output of function blockfile()}
}
\value{
\code{maskfile()} also returns the following list of objects:
\item{file}{list characters; indicating filename and root folder where text file is stored}
\item{zones}{list list of values to be passed to ssdpars()}
}
\description{
Creates a mask file and writes the result into a text file (.out) that is stored in \code{input} folder.
The text file (.out) stores values {-1,0} where -1 are assigned to nodata locations and 0 are assigned to nodes with values (id region).
As input you should provide the name of list returned by funtion blockfile().
}
