% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{varmodel}
\alias{varmodel}
\title{Fit a theoretical semi-variogram to data}
\usage{
varmodel(varexp, mod = c("exp", "sph"), nug, ran, sill)
}
\arguments{
\item{varexp, }{string, name of object, output of function `varexp()`}

\item{mod, }{character, the semi-variogram model type ('sph' or 'exp')}

\item{nug, }{numeric, nugget-effect value of the semi-variogram}

\item{ran, }{numeric, range value of the semi-variogram}

\item{sill, }{numeric, sill (or partial sill) value of the semi-variogram}
}
\value{
Function returns the following list of objects:
\item{structures}{the number of spatial strutures (excluding nugget-effect)}
\item{parameters}{a dataframe with model information to be passed to ssdpars()}
\item{fittedval}{a numeric vector of fitted values as set by the semi-variogram model}
}
\description{
Function fits (manually) a theoretical semi-variogram.
For now, model types available are spherical ('sph') and exponential ('exp').
Users should provide the experimental semi-variogram, fit a model type and set the semi-variogram parameters (nugget, range and sill).
With base function plot(), results can be visualized.
}
