% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptgrid}
\alias{ptgrid}
\title{ptgrid dataset}
\format{
A data.frame with 40608 rows and 3 columns
\describe{
  \item{x}{x-coordinates, in meters}
  \item{y}{y-coordinates, in meters}
  \item{oid_}{id, municipality's id number)}
}
}
\source{
Data preparation by Manuel Ribeiro & Leonardo Azevedo (leonardo.azevedo@tecnico.ulisboa.pt).
}
\usage{
data(ptgrid)
}
\description{
The ptgrid dataset refers to a data.frame representing
a rectangular grid with 2 km x 2 km spacing that covers Portugal mainland.
Grid nodes with real values refer to the municipality id (block data).
Grid nodes located on sea or Spain country are set to `NA`.
}
\examples{
data(ptgrid)
str(ptgrid)
}
\references{
Azevedo, L. et al. Geostatistical COVID-19 infection risk maps for Portugal. Int J Health Geogr 19, 25 (2020).
}
\keyword{datasets}
