% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{pxmap}
\alias{pxmap}
\title{Creates a pixelated map}
\usage{
pxmap(
  mapobj = maps,
  nbigk = c(17, 17),
  bigk = 4,
  scaleft = 1,
  legname = "Median\\nCum. risk"
)
}
\arguments{
\item{mapobj}{string, name of list, output of function `outraster()`}

\item{nbigk}{vector with length 2, specifies the minimum number of large pixels in the x and y directions, respectively.}

\item{bigk}{integer, specifies the number of average quantile intervals (i.e. number of different pixel sizes).}

\item{scaleft}{integer, specifies a factor (in units of observations) that features in either iterative multiplication or iterative exponentiation (see `scale_factor` arg in `pixelate::pixelate() for more details).}

\item{legname}{character, legend title.}
}
\value{
`pxmap()` is used to map disease risk with varying pixel sizes
showing spatial uncertainty generated by
block direct sequential simulation algorithm.
}
\description{
This is a wrapper function calling R package `pixelate`^[https://github.com/aimeertaylor/pixelate]
developed by Aimee Taylor and colleagues ^[https://doi.org/10.48550/arXiv.2005.11993] providing tools
to plot disease risk mapping with a visual representation of spatial uncertainty,
as a function of pixel size.
}
\details{
As input you should provide the result returned by `outraster()` (a list of `RasterLayer` objects)
and additional arguments required for pixelation.
}
