% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{spmap}
\alias{spmap}
\title{Creates elegant median e-type/uncertainty maps}
\usage{
spmap(mapobj = maps, mapvar = c("etype", "unc"), legname = "Incidence")
}
\arguments{
\item{mapobj}{string, name of list, output of function `outraster()`}

\item{mapvar}{character, specifies the minimum number of large pixels in the x and y directions, respectively.}

\item{legname}{character, specifies the number of average quantile intervals (i.e. number of different pixel sizes).}
}
\value{
`spmap()` is used to map a spatially continuous disease risk
or spatial uncertainty generated by the block direct sequential simulation algorithm.
}
\description{
The function wraps functions from ggplot2 to provide a more elegant map.
}
\details{
As input you should provide the result returned by `outraster()` or a `RasterLayer` object.
}
