% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{spmap}
\alias{spmap}
\title{Creates elegant simulations, median e-type/uncertainty maps}
\usage{
spmap(mapobj = maps, mapvar = "etype", simid = 1, legname = NA)
}
\arguments{
\item{mapobj}{string, name of list, output of function `outraster()`}

\item{mapvar}{character, the map to be plotted, one of the following c("etype", "unc", "simulations")}

\item{simid}{numeric, for mapvar = "simulations", is the simulation number to be plotted.}

\item{legname}{character, legend title.}
}
\value{
`spmap()` is used to map spatially continuous maps representing disease risk simulations, median e-type disease risk
or spatial uncertainty (inter-quartile difference) generated by the block direct sequential simulation algorithm.
}
\description{
The function wraps functions from ggplot2 to provide a more elegant map.
}
\details{
As input you should provide a `RasterLayer` object returned by `outraster()`.
}
